Add files via upload
In this project, we use the well-known IMDB dataset, which comprises 25,000 movie reviews classified as either positive or negative. The objective is to train a machine learning model to accurately predict the sentiment of each review.

The approach involves training the model to determine whether a review is overall positive or negative. Instead of using traditional word embeddings, we utilize the TextVectorization layer for preprocessing the text data. This method transforms the text into a format suitable for model training, simplifying the process of feature extraction and improving the efficiency of the training pipeline.